Move PublicationFieldValue up to parent namespace
